api side fixes

+++ api/migrations/20240201142318_create_locations.up.sql
@@ -1,10 +1,10 @@
 create table locations(
     id varchar(63) not null primary key,
 
-    lat double not null,
-    lng double not null,
-    alt double not null,
-    precision double not null,
+    lat double precision not null,
+    lng double precision not null,
+    alt double precision not null,
+    precision double precision not null,
 
     user_id varchar(63) not null references users(id),
